requirements-backend.txt
fastapi>=0.110
uvicorn[standard]>=0.23
SQLAlchemy>=2.0
PyJWT>=2.8
bcrypt>=4.0
python-dotenv>=1.0
httpx>=0.27
python-dateutil>=2.8
PyMySQL>=1.1
python-multipart>=0.0.9
pydantic>=2.5
email-validator>=2.1
